actors/pyproject.toml
[tool.poetry]
name = "askgpt-actors"
version = "0.1.0"
description = "Distributed LLM inference for askgpt"
authors = ["raceychan <raceychan@gmail.com>"]
license = "GNU"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ray = { version = "2.9.0", extras = ["serve"] }
grpcio-tools = "1.60.0"
llama-cpp-python = "^0.2.25"
pure-protobuf = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''(*_pb2.py | *_pb2_grpc.py)'''

[tool.isort]
skip = ["*_pb2.py", "*_pb2_grpc.py"]
